+++ azure-storage-blob/azure/storage/blob/baseblobservice.py
@@ -63,6 +63,7 @@ from ._deserialization import (
     _parse_lease,
     _convert_xml_to_signed_identifiers_and_access,
     _parse_base_properties,
+    _parse_account_information,
 )
 from ._download_chunking import _download_blob_chunks
 from ._error import (
@@ -1320,6 +1321,33 @@ class BaseBlobService(StorageClient):
 
         return self._perform_request(request, _convert_xml_to_blob_list, operation_context=_context)
 
+    def get_blob_account_information(self, container_name=None, blob_name=None, timeout=None):
+        """
+        Gets information related to the storage account.
+        The information can also be retrieved if the user has a SAS to a container or blob.
+
+        :param str container_name:
+            Name of existing container.
+            Optional, unless using a SAS token to a specific container or blob, in which case it's required.
+        :param str blob_name:
+            Name of existing blob.
+            Optional, unless using a SAS token to a specific blob, in which case it's required.
+        :param int timeout:
+            The timeout parameter is expressed in seconds.
+        :return: The :class:`~azure.storage.blob.models.AccountInformation`.
+        """
+        request = HTTPRequest()
+        request.method = 'HEAD'
+        request.host_locations = self._get_host_locations(secondary=True)
+        request.path = _get_path(container_name, blob_name)
+        request.query = {
+            'restype': 'account',
+            'comp': 'properties',
+            'timeout': _int_to_str(timeout),
+        }
+
+        return self._perform_request(request, _parse_account_information)
+
     def get_blob_service_stats(self, timeout=None):
         '''
         Retrieves statistics related to replication for the Blob service. It is 
@@ -1358,7 +1386,7 @@ class BaseBlobService(StorageClient):
 
     def set_blob_service_properties(
             self, logging=None, hour_metrics=None, minute_metrics=None,
-            cors=None, target_version=None, timeout=None, delete_retention_policy=None):
+            cors=None, target_version=None, timeout=None, delete_retention_policy=None, static_website=None):
         '''
         Sets the properties of a storage account's Blob service, including
         Azure Storage Analytics. If an element (ex Logging) is left as None, the 
@@ -1393,6 +1421,11 @@ class BaseBlobService(StorageClient):
             It also specifies the number of days and versions of blob to keep.
         :type delete_retention_policy:
             :class:`~azure.storage.common.models.DeleteRetentionPolicy`
+        :param static_website:
+            Specifies whether the static website feature is enabled,
+            and if yes, indicates the index document and 404 error document to use.
+        :type static_website:
+            :class:`~azure.storage.common.models.StaticWebsite`
         '''
         request = HTTPRequest()
         request.method = 'PUT'
@@ -1405,7 +1438,7 @@ class BaseBlobService(StorageClient):
         }
         request.body = _get_request_body(
             _convert_service_properties_to_xml(logging, hour_metrics, minute_metrics,
-                                               cors, target_version, delete_retention_policy))
+                                               cors, target_version, delete_retention_policy, static_website))
 
         self._perform_request(request)